+++ out/test/suite/testFiles/123.js
@@ -1,0 +1,2 @@
+"use strict";
+//# sourceMappingURL=123.js.map